// client/src/types/cat.ts
export interface CatBreed {
  _id: string;
  name: string;
  imageURLs: any;
  origin: string;
  country_code: string;
  alt_names: string;
  life_span: string;
  temperament: any;
  wikipedia_url: string;
  weight: any;
  height: any;
  adaptability: number;
  affection_level: number;
  child_friendly: number;
  dog_friendly: number;
  grooming: number;
  intelligence: number;
  health_issues: number;
  social_needs: number;
  stranger_friendly: number;
  vocalisation: number;
  energy_level: number;
  hypoallergenic: number;
  shedding_level: number;
  description: string;
}